Added spectrum grapher to repo, added ability to read multi-spectrum files with data_process
I've been using spectrum_grapher for a while now, but it's pretty good and it allows for adding in new features as I need them (though I do need to organize it a bit better so it doesn't become a huge mess). I thought I should add it to the repo since there's no reason it can't run on other computers (and it's like super useful).

Multi-spectrum loading and saving is kind of janky at the moment, but it's good enough for now, and it does work well enough.

Later on, I might try integrating spectrum_grapher with data_process to make it so I only need one tool for all the data analysis.